# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个游戏管理系统的API文档，包含前端用户、后台管理、游戏管理等功能",
        "title": "游戏管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8801",
    "basePath": "/",
    "paths": {
...
        "/v1/admin/login": {
            "post": {
                "description": "管理员使用账号密码登录系统，返回JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-管理员认证"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员登出系统，清除token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-管理员认证"
                ],
                "summary": "管理员登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/admin/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的管理员账号（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-管理员认证"
                ],
                "summary": "创建管理员账号",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
        "dto.AdminLoginReq": {
            "type": "object",
            "required": [
                "admin_name",
                "password"
            ],
            "properties": {
                "admin_name": {
                    "description": "管理员账号",
                    "type": "string",
                    "example": "admin"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "dto.AdminRegisterReq": {
            "type": "object",
            "required": [
                "admin_name",
                "email",
                "password",
                "real_name"
            ],
            "properties": {
                "admin_name": {
                    "description": "管理员账号",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3,
                    "example": "admin"
                },
                "department": {
                    "description": "部门",
                    "type": "string",
                    "maxLength": 50,
                    "example": "技术部"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "admin@example.com"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6,
                    "example": "123456"
                },
                "phone_number": {
                    "description": "手机号",
                    "type": "string",
                    "example": "13800138000"
                },
                "position": {
                    "description": "职位",
                    "type": "string",
                    "maxLength": 50,
                    "example": "系统管理员"
                },
                "real_name": {
                    "description": "真实姓名",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2,
                    "example": "张三"
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 500,
                    "example": "系统管理员账号"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}